Simplified and cleaned up [PLEASE READ DESCRIPTION]
Found a missing "{", removed the else if statements because else if skips other all other else if statements if it is done, used if instead. Also changed the correction tool so that the position can't go over 1 or under 0 code is put into the if statement instead of being a separate if statement within the other one.